Create README.md
Most clients would prefer to settle confidentially, out of court. But the quickest path to a fair out-of-court settlement is to hire a lawyer with a reputation for going to trial and winning. That’s because lawyers without a trial record aren’t taken seriously by large corporations or insurance companies.